feat: high performance read dart entities from cpp.
Former-commit-id: 46c7863e5738f660e6f0b5707a49da705af620ee [formerly f76880ed72ee3b89d6110ae6d8b4ce8e384a8731]
Former-commit-id: c4def09c3e2bbc4ce75cfb54eafb8f5b257a53ae

+++ kraken/lib/widget.dart
@@ -130,7 +130,7 @@ class KrakenRenderWidget extends SingleChildRenderObjectWidget {
 
     if (kProfileMode) {
       PerformanceTiming.instance(controller.view.contextId)
-          .mark(PERF_CONTROLLER_INIT_START, controllerCreateStart.microsecondsSinceEpoch.toDouble());
+          .mark(PERF_CONTROLLER_INIT_START, controllerCreateStart.microsecondsSinceEpoch);
       PerformanceTiming.instance(controller.view.contextId).mark(PERF_CONTROLLER_INIT_END);
     }
